back-end/requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
flask_cors==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==3.1.0
greenlet==3.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
flask-wtf
